happier linters

+++ all.bash
@@ -1,18 +1,27 @@
 #!/bin/bash
-set -e
+set -eu
 
+# setup
+cd "$( dirname "${BASH_SOURCE[0]}" )"
+export GOBIN=${GOBIN-$PWD/build}
+export PATH=$PWD/build:$PATH
 cmd=${*-test}
 protoc_flags=(
   -I=.
+  -I=$PWD/vendor/github.com/gogo/protobuf/protobuf
   -I=$GOPATH/src
-  -I=$GOPATH/src/github.com/gogo/protobuf/protobuf
   --gogoslick_out=Mgoogle/protobuf/timestamp.proto=github.com/gogo/protobuf/types:./
 )
 protoc_flags="${protoc_flags[@]}"
 
+# check / install dependencies
+if ! which protoc >/dev/null ; then
+  echo 'Please install external dependency: protobuf (protoc program)' >&2
+  exit 1
+fi
+
 go get \
   github.com/alecthomas/gometalinter \
-  github.com/kovetskiy/manul \
   github.com/gogo/protobuf/proto \
   github.com/gogo/protobuf/protoc-gen-gogoslick \
   github.com/gogo/protobuf/gogoproto
@@ -23,7 +32,9 @@ patch -d ./vendor/github.com/fiorix/go-smpp --forward -r- -p1 --batch <patch/go-
 
 go get -tags trace ./vendor/github.com/temoto/go-sqlite3
 
+# essence
 ( cd talk ; protoc $protoc_flags *.proto )
+[[ "$cmd" == "proto" ]] && exit 0
 
 if [[ "$cmd" == "test" ]] ; then
   rm -f coverage.txt
@@ -33,8 +44,7 @@ if [[ "$cmd" == "test" ]] ; then
     [[ -f "$tmp" ]] && cat $tmp >>coverage.txt
   done
 elif [[ "$cmd" == "build" ]] ; then
-  go build -tags trace ./...
-  GOBIN=$PWD go get -tags trace ./...
+  go get -tags trace ./...
 else
   go $cmd ./...
 fi
